+++ packages/storybook/.storybook/preview.js
@@ -105,5 +105,5 @@ configureActions({
     depth: 1
 });
 
-// Update the GUID on this line to trigger a turbosnap full rebuild: 123e4567-e89b-12d3-a456-426614174000
+// Update the GUID on this line to trigger a turbosnap full rebuild: 300cddbf-b808-403a-8561-9e41eed8a155
 // See https://www.chromatic.com/docs/turbosnap/#full-rebuilds