+++ packages/angular-workspace/nimble-angular/CHANGELOG.md
@@ -1,8 +1,48 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Fri, 02 Aug 2024 22:55:10 GMT and should not be manually modified.
+This log was last generated on Mon, 12 Aug 2024 21:44:15 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 27.0.6
+
+Mon, 12 Aug 2024 21:44:15 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v31.2.0
+
+## 27.0.5
+
+Sat, 10 Aug 2024 00:15:34 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v31.1.4
+
+## 27.0.4
+
+Fri, 09 Aug 2024 15:08:21 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v31.1.3
+
+## 27.0.3
+
+Wed, 07 Aug 2024 16:58:22 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v31.1.2
+
+## 27.0.2
+
+Tue, 06 Aug 2024 16:12:52 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v31.1.1
 
 ## 27.0.1